fix(example): exports ffi for test usages
