+++ src/api/eth/proposal.rs
@@ -9,7 +9,6 @@ use actix_web::{web, Responder};
 use cmmr::MMR;
 use reqwest::Client;
 use scale::Decode;
-use std::{ffi::CStr, os::raw::c_char};
 
 /// Proposal post req
 #[derive(Deserialize)]
@@ -29,16 +28,12 @@ impl ProposalReq {
 
     /// Get `EtHashProof`
     fn ethash_proof(block: u64) -> Vec<EthashProofJson> {
-        unsafe {
-            let proof = CStr::from_ptr(Proof(block as u32))
-                .to_string_lossy()
-                .to_string();
-            <Vec<EthashProof>>::decode(&mut bytes!(proof.as_str()).as_ref())
-                .unwrap_or_default()
-                .iter()
-                .map(Into::<EthashProofJson>::into)
-                .collect()
-        }
+        let proof = super::ffi::proof(block);
+        <Vec<EthashProof>>::decode(&mut bytes!(proof.as_str()).as_ref())
+            .unwrap_or_default()
+            .iter()
+            .map(Into::<EthashProofJson>::into)
+            .collect()
     }
 
     // Get mmr root
@@ -98,11 +93,7 @@ pub struct ProposalHeader {
     mmr_proof: Vec<String>,
 }
 
+/// Proposal Handler
 pub async fn handle(req: web::Json<ProposalReq>) -> impl Responder {
     web::Json(req.0.headers().await)
-}
-
-#[link(name = "darwinia_shadow")]
-extern "C" {
-    fn Proof(input: libc::c_uint) -> *const c_char;
 }
